For each phenological record (day+year) at every observational and experimental site, I calculated a) the cumulative base zero GDDs from Jan 1 to the phone date, b) cumulative base 5 GDDs from Jan 1 to the phone date, and c) chill days (# days mean temp <5 C) from the previous October-December of the phone date, all using the GLDAS data that Christy put together. These calculations are saved (one file per site) in /Analyses/teambackground/output/obsclim for the observational sites and /Analyses/teambackground/output/expclim for the experimental sites.
We’re missing calculations for a bunch of the experimental data because a lot of the experiments were post 2010, after the GLDAS data is unavailable. We’re also missing data from any non-North American sites because we don’t have the GLDAS data for that yet. That said, the code is easy and quick to rerun, so if we get updated or new climate datasets formatted the same as the GLDAS, I can easily rerun things.