Warming Effects on Temperature & Moisture Figure Updates
Updated (potential) manuscript figures for the whole group paper.  All
figures were re-generated to reflect corrected file path.  Also change
the coding scheme to be soft-coded a bit better.

There were two major changes to figure content to reflect requests by
Ailene in draft #2:
1) added mean annual temperature (or soil moisture) as text to
[…]_Deviation.png figures.
2) adjust mean lines to show range of differences in observed
temperature or moisture in […]_AnnMeans.png figures.

In all cases, mean annual temperature/moisture was calculated by first
averaging across years for each day of year and then I took the mean
across years.  This was done because few to any sites had consistent
coverage for all days of the year (as evidenced by the wiggly mean
lines) and this seemed to be the best option.  Can be redone to average
across days and then across years, which should produce a very similar
answer, but possibly not identical.

For the […]_Deviation[…] figures, we first aggregate across years for
each plot for each day of year.  We then use that plot/day-level to
calculate the mean line and 95% CIs from the variability seen across
plots.  Thus the 95% ci shown around treatments and the “baseline”
conditions indicates the variability among replicates.