Slight figure tweaks per emails with Ailene
excerpt of email from Ailene:
1) Make Figure 2 a 2 x 5 figure, and order plots by MAT. For exp 12, it
looks like there is a large gap in datacollection- could you make this
appear as a gap, instead of as continuous values?
2) Make Figure 4 a 2 X 5 figure, and order plots by mean soil moisture
(add this to the plots, in the same way that you added MAT in Figure
2). Also, I’m thinking we should probably remove exp08 from the soil
moisture figure, since it is such a short time period. What do you
think?